Adadpt to CCS and fix bug in matrix computation

+++ modelEvaluation/compare_sex_riskfactors.py
@@ -66,9 +66,15 @@ if not os.path.exists(filename):
     sex=['Mujeres', 'Hombres']
     
     X.drop('PATIENT_ID', axis=1, inplace=True)
+    
+    # cols = {c:X[c]*X['FEMALE'] for c in X.drop('FEMALE',axis=1).columns}
+    
+    
+    # out = pd.concat([ df.astype(np.int8).add_suffix(f'{k}INTsex') for k,df in cols.items()], axis=1)
+    # assert False
     for column in X:
         if column!='FEMALE':
-            X[f'{column}INTsex']=X[column]*X['FEMALE']
+            X[f'{column}INTsex']=(X[column]*X['FEMALE']).astype(np.int8)
     features=X.columns
     stderrInt, zInt, pInt=beta_std_error(model, X)
     print('Intercept: ',list(model.intercept_))
